Debug the project

+++ order-service/src/main/java/org/example/orders/repository/OrderRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface OrderRepository {
     void save(OrderEntity order);
-    OrderEntity findById(Long id);
+    OrderEntity findById(Integer id);
     List<OrderEntity> findByDateAndSum(Date createDate, Long sum, Long scale);
     List<OrderEntity> findWithoutOrderAndBetweenDate(String number, Date dateBefore, Date dateAfter);
 }
